client-go: remove import of github.com/gregjones/httpcache
